// src/types/ChatGPTPlanType.ts
export interface ChatGPTPlanType {
    city: string;
    day: string;
    travel_destination_1: string;
    travel_destination_1_local: string;
    travel_destination_1_time: string;
    travel_destination_1_detail: string;
    travel_destination_2: string;
    travel_destination_2_local: string;
    travel_destination_2_time: string;
    travel_destination_2_detail: string;
    travel_destination_3: string;
    travel_destination_3_local: string;
    travel_destination_3_time: string;
    travel_destination_3_detail: string;
    restaurant_1: string;
    restaurant_1_local: string;
    restaurant_1_time: string;
    restaurant_1_detail: string;
    restaurant_2: string;
    restaurant_2_local: string;
    restaurant_2_time: string;
    restaurant_2_detail: string;
    restaurant_3: string;
    restaurant_3_local: string;
    restaurant_3_time: string;
    restaurant_3_detail: string;
}